Rearrange vim config files
This change drops the scheme where vimrc is hub for configurations.
Instead vimrc holds the actual configuration. If I need to have second
configuration I can use an alias with `vim -u /path/to/config`. In
addition if I need to change the `.vim` directory I can do it with `set
runtimepath`.

Reads:
https://superuser.com/questions/209075/toggle-between-two-different-vim-configurations
https://www.mojotech.com/blog/vimrc-tutorial/