Add shell init files
Prior to this change there was no shell init file configuration. The
configuration sets my preferred configuration for Linux and MacOS.

~/.inputrc is where readline custom settings are stored and read by bash
and other programs that use readline.

~/.bash_profile can be used instead of ~/.profile, but it is read by
bash only, not by any other shell. The file is here because MacOS use
the file. It is configured to source ~/.profile.

~/.profile is the place to put stuff that applies to your whole session,
such as programs that you want to start when you log in , and
environment variable definitions that impact not only bash sessions.

~/.bashrc is the place to put stuff that applies only to interactive
bash shells, such as alias and function definitions, shell
options, and prompt settings.

~/.bash_aliases is the place to set aliases.

~/.bashrc_macos is only for MacOS specific settings.

~/.bashrc_linux is only for Linux specific settings.

Reads:
https://superuser.com/questions/183870/difference-between-bashrc-and-bash-profile
https://www.tecmint.com/understanding-shell-initialization-files-and-user-profiles-linux/
http://www.joshstaiger.org/archives/2005/07/bash_profile_vs.html